fixed formatting, changed from pico-sdk style to an object oriented style
more generalized for other microcontrollers

+++ include/dma.h
@@ -2,14 +2,15 @@
 #define DMA_H
 
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 
 int dma_config(bool high_priority);
 
-void dma_start_transfer(int chan, const void *src, void *dst, uint32_t count);
-bool dma_is_busy(int chan);
-void dma_wait_complete(int chan);
+void dma_start_transfer(uint8_t chan, const void *src, void *dst, uint32_t count);
+bool dma_is_busy(uint8_t chan);
+void dma_wait_complete(uint8_t chan);
 
-void dma_configure_peripheral(int chan, const volatile void *src, volatile void *dst, uint32_t count, bool read_from_peripheral);
+void dma_configure_peripheral(uint8_t chan, const volatile void *src, volatile void *dst, size_t count, bool read_from_peripheral);
 
 #endif
